Adds Warehouse page and mods old Warehouse to be Warehouses.
