+++ src/pages/Warehouse/warehouse.scss
@@ -1,34 +1,2 @@
 @use "../../styles/partials/variables.scss" as *;
 @use "../../styles/partials/mixins.scss" as *;
-
-.warehouse {
-    position: relative;
-
-    &__page-background {
-        background-color: $background-graphite;
-        position: absolute;
-        width: 100%;
-        height: 3.75rem;
-        z-index: -1;
-
-        @include tablet-and-desktop {
-            height: 7rem;
-        }
-    }
-
-    &__page-foreground {
-        margin-left: $m-site-wide-left-right-padding;
-        margin-right: $m-site-wide-left-right-padding;
-        background-color: $primary-instock-white;
-
-        @include tablet {
-            margin-left: $t-site-wide-left-right-padding;
-            margin-right: $t-site-wide-left-right-padding;
-        }
-    
-        @include desktop {
-            margin-left: $site-wide-left-right-padding;
-            margin-right: $site-wide-left-right-padding;
-        }
-    }
-}
